feat: criação do metodo isToxic dentro do serviço moderateService
